fix: add folder for components & style, NavbarUser early return & style, file naming
